Revert some changes to intro changes
