+++ src/client/components/AppContainer.tsx
@@ -1,55 +1,10 @@
-import { Dispatch, PropsWithChildren, SetStateAction, useContext, useState } from 'react'
-import { Nav, NavLink } from 'react-bootstrap'
+import { PropsWithChildren, useState } from 'react'
 import Container from 'react-bootstrap/esm/Container'
-import Form from 'react-bootstrap/esm/Form'
-import Navbar from 'react-bootstrap/esm/Navbar'
-import NavbarCollapse from 'react-bootstrap/esm/NavbarCollapse'
 import Stack from 'react-bootstrap/esm/Stack'
 import { useProjects } from '../projects'
-import { Branding } from './Branding'
 import { CircularProgressIndicator } from './CircularProgressIndicator'
-import { IntroContext } from './IntroPage'
+import { Header } from './Header'
 import { ProjectList } from './ProjectList'
-
-const IntroNavLink = () => {
-  const [_show, setShow] = useContext(IntroContext)
-
-  const onClick = () => setShow(true)
-
-  return <NavLink onClick={onClick}>Intro</NavLink>
-}
-
-
-interface HeaderProps {
-  setSearch: Dispatch<SetStateAction<string>>
-}
-
-const Header = ({ setSearch }: HeaderProps) => (
-  <Navbar expand='lg' className='navbar-main bg-body-tertiary' sticky='top'>
-    <Container>
-      <Navbar.Brand href='/' className='col-xs-8 col-sm-4 d-flex justify-content-start align-items-stretch flex-shrink'>
-        <Branding />
-      </Navbar.Brand>
-      <Navbar.Toggle />
-      <NavbarCollapse>
-        <Nav className='d-flex justify-content-end flex-fill ms-xl-4 fs-5'>
-          <NavLink href='/projects'>Projects</NavLink>
-          <IntroNavLink/>
-          <NavLink href='/help'>Help</NavLink>
-          <NavLink href='/about'>About</NavLink>
-          <Form className='ms-lg-4 ms-xxl-5 me-2 search'>
-            <Form.Control
-              type='search'
-              placeholder='Search'
-              aria-label='Search'
-              onChange={e => setSearch(e.target.value)}
-            />
-          </Form>
-        </Nav>
-      </NavbarCollapse>
-    </Container>
-  </Navbar>
-)
 
 interface SearchProjectListProps {
   search: string
