--- src/client/components/AuthorPage.tsx
@@ -1,32 +1,0 @@
-import { useProjects } from '../projects'
-import { ProjectInfo } from '../types'
-import { AppContainer } from './AppContainer'
-import { CircularProgressIndicator } from './CircularProgressIndicator'
-import { FilteredProjectListContainer } from './FilteredProjectList'
-
-interface AuthorProjectsProps {
-  name: string
-  projects: ProjectInfo[]
-}
-
-const AuthorProjects = ({ name, projects }: AuthorProjectsProps) => {
-  const filteredProjects = projects.filter(e => e.authors.includes(name))
-  return <FilteredProjectListContainer projects={filteredProjects} />
-}
-
-interface AuthorPageProps {
-  name: string
-}
-
-export const AuthorPage = (props: AuthorPageProps) => {
-  const projects = useProjects()
-  if (projects == undefined) {
-    return (
-      <AppContainer>
-        <CircularProgressIndicator />
-      </AppContainer>
-    )
-  }
-
-  return <AuthorProjects {...props} projects={projects} />
-}
