+++ src/client/components/Branding.tsx
@@ -6,7 +6,7 @@ const Logo = () => {
       height='65'
       width='50'
       className='siteLogo d-none d-md-inline'
-   >
+    >
       <path d='m23.41,13.51c-.5-1.6-1.3-2.4-2.8-3.2-1.5-.9-3.6-.6-5,.3-.4.2-.7.5-1.1.8-.1,0-.3.2-.5.3v-.2c0-.3,0-.6-.1-.9l-.1.3.2-.5c.2-.6.1-1,0-1.5-.2-.6-.1-.8,0-1.4h0c-.5-.1.9-1,1.3-1,.6.1,1.1,0,1.7,0,1.3-.4,2.6-.7,4-.7,1,0,2.6-1.9,2.7-3-1.3.1-2.5.2-3.8.3-1.2.2-2.4.2-3.7.3-.3.1-.6.1-.9.1s-.6.1-.9.2c-.5.1-.5,0-.3-.2,0-.3.1-.7.1-1s.1-.5.1-.8c0-.4.1-.8-.2-.1.3-.7.5-1.7-.6-1.6-.9,0-2.1,1.2-2.4,2-.2.6-.3,1-.3,1.6,0,0,.1,0,.1.1.4.2-1.1,1.2-1.5.9-.9-.6-1.8-.3-2.8-.2-.8.1-1.6,0-2.5,0s-1.9-.1-2.8-.1C.51,5.01-.09,6.11.01,6.81c2.3-.3,4.5-.1,6.8-.1,1.2,0,3.5-.5,4.4.5.1.1-.2.4-.7.7v1.9c-.1.9,0,1.4.2,2.2-.2,1.1-.1,2.3-.2,3.5,0,1.4,1.6.6,2.1.1q.1-.1.2-.1c.1,0,.2-.1.4-.2.2-.1.4-.2.5-.3.5-.3,1-.7,1.5-1.1,1.1-.9,2.3-.3,3.4.4,1.1.7,1.3,1.5,1.7,2.7.1.2.2,1.4.4,1.8-.1.1-.2.1-.2.2-.3.3-.5.6-.8.9.1-.1-.3.1-.4.2-.6.5-.9,1-1.4,1.6-.2.3-.5.6-.7.8,0,0-.6.8-.5.6l-1.5,1.5c.3-.3-.2.2-.4.3l-.1.1c-.3.5-.7.9-1.2,1.3-1,.9-2,1.6-2.9,2.6-.8.9-1.7,1.9-2.6,2.7h.1c.5.3,1.3,0,1.9-.5.5-.3,1.1-.5,1.5-.8.9-.5,1.7-1.2,2.5-1.9.1-.1.2-.2.3-.3.3-.3.8-.5,1.1-.7.7-.6,1.3-1.3,2-2,.5-.6,1.1-1,1.6-1.7.3-.4.6-.7.8-1.1.1-.1.2-.3.4-.5,0,0,.1,0,.1-.1.3-.3.7-.5.9-.8l.2-.2c.2-.1.4-.3.6-.4.8-.7,1.4-1.5,1.8-2.5.7-1.5.1-3.1-.4-4.6' />
     </svg>
   )
