+++ src/client/main.tsx
@@ -5,18 +5,13 @@ import '../scss/styles.scss'
 import { AboutPage } from './components/AboutPage'
 import { HelpPage } from './components/HelpPage'
 import { HomePage } from './components/HomePage'
-import { IntroPage, IntroContext, useIntro } from './components/IntroPage'
-{/*import {SitePage} from './components/SitePage'*/ }
+import { IntroContext, IntroPage, useIntro } from './components/IntroPage'
+import { ProjectsPage } from './components/ProjectsPage'
 
 const Router = () => <>
   <Route path='/'><HomePage /></Route>
-  <Route path='/projects'><HomePage /></Route>
   <Route path='/intro'><IntroPage /></Route>
-  {/*<Route path='/project/:project'>{({ project }) => <ProjectPage name={decodeURI(project)} />}</Route>*/}
-  {/*<Route path='/author/:author'>{({ author }) => <AuthorPage name={decodeURI(author)} />}</Route>*/}
-  {/*<Route path='/tag/:tag'>{({ tag }) => <TagPage tag={decodeURI(tag)} />}</Route>*/}
-  {/*<Route path='/status/:status'>{({ status }) => <StatusPage status={decodeURI(status)} />}</Route>*/}
-  {/*<Route path='/site/:site'>{({site}) => <SitePage site={decodeURI(site)}/>}</Route>*/}
+  <Route path='/projects'><ProjectsPage /></Route>
   <Route path='/about'>{<AboutPage />}</Route>
   <Route path='/help'>{<HelpPage />}</Route>
 </>
@@ -26,7 +21,6 @@ const App = () => {
   return (
     <IntroContext.Provider value={[showIntro, setShowIntro]}>
       <Router />
-      <IntroPage />
     </IntroContext.Provider>
   )
 }
